Move BoxOfficeServe into example.adapter.command package
