+++ sample-app/src/test/scala/example/application/command/MyBoxOfficeServiceSpec.scala
@@ -1,7 +1,7 @@
 package example.application.command
 
 import akka.actor.typed.ActorSystem
-import example.adapter.BoxOfficeService
+import example.adapter.command.BoxOfficeService
 import example.application.ApplicationDIDesign
 import example.testing.tags.ExerciseTest
 import testkit.AirframeDiSessionSupport
